src/config.py
from __future__ import annotations

import os
from dataclasses import dataclass

from dotenv import load_dotenv

load_dotenv()


def _to_bool(value: str | None, default: bool = False) -> bool:
    if value is None:
        return default
    return value.strip().lower() in {"1", "true", "yes", "on"}


@dataclass(frozen=True)
class Settings:
    risk_approval_threshold: float = float(os.getenv("RISK_APPROVAL_THRESHOLD", "0.7"))

    llm_enabled: bool = _to_bool(os.getenv("LLM_ENABLED", "true"), default=True)
    groq_api_key: str = os.getenv("GROQ_API_KEY", "")
    groq_model: str = os.getenv("GROQ_MODEL", "llama-3.3-70b-versatile")
    groq_temperature: float = float(os.getenv("GROQ_TEMPERATURE", "0.2"))


settings = Settings()