pyproject.toml
[project]
name = "agentic-pr-pipeline"
version = "0.1.0"
description = "Zero-to-Production Agentic PR Pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7,<3.0",
  "langgraph>=0.2.40,<0.7.0",
  "langchain>=0.3.21,<0.4.0",
  "langchain-core>=0.3.49,<0.4.0",
  "langchain-groq>=0.2.4,<0.3.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "fastapi>=0.115.11,<1.0.0",
  "uvicorn>=0.34.0,<1.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "httpx>=0.28.1,<1.0.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]